Commit as at 2024-09-19 2:23pm. Update of Charts

+++ src/main/java/com/ngxgroup/xticket/payload/ChartPayload.java
@@ -15,7 +15,7 @@ import lombok.Setter;
 @NoArgsConstructor
 public class ChartPayload {
 
-    int[] data;
+    long[] data;
     String name;
     String type;
 }
